% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_check.R
\name{gp_check}
\alias{gp_check}
\title{Wrapper of goodpractice gp}
\usage{
gp_check(path = ".", checks = all_checkers(), extra_preps = all_prepers(),
  extra_checks = all_extra_checkers(), quiet = TRUE)
}
\arguments{
\item{path}{Path to a data analysis root.}

\item{checks}{Character vector, the checks to run. Defaults to
all checks.}

\item{extra_preps}{Custom preparation functions. See
\code{\link[goodpractice]{make_prep}} on creating preparation functions.}

\item{extra_checks}{Custom checks.}

\item{quiet}{Whether to suppress output from the preparation
functions. Note that not all preparation functions produce output,
even if this option is set to \code{FALSE}.}
}
\value{
A checkers object that you can query
  with a simple API. See \code{\link{results}} to start.
}
\description{
Wrapper of goodpractice gp
}
\examples{
check_results <- gp_check(path=system.file("scripts", package="checkers"),
         checks = "comments",
         extra_preps = list(scripts = prep_scripts),
         extra_checks = list(comments = check_well_commented))
check_results
}
