% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\name{load_config}
\alias{load_config}
\title{Load checkers config}
\usage{
load_config(filename)
}
\arguments{
\item{filename}{string to custom file}
}
\description{
Loads the config file into options which are
used elsewhere in the application.
}
\examples{
load_config()
}
