% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lintr.R
\name{prep_lint_dir}
\alias{prep_lint_dir}
\title{Prep for linting
This prep function is used to overwrite the goodpractice "lintr" prep
function so as to lint the entire directory rather than just package
files.}
\usage{
prep_lint_dir(state, path = state$path, quiet)
}
\description{
Prep for linting
This prep function is used to overwrite the goodpractice "lintr" prep
function so as to lint the entire directory rather than just package
files.
}
